INT-3338: Add `MongoDbChannelMessageStore`
