+++ spring-integration-mongodb/src/main/java/org/springframework/integration/mongodb/store/ConfigurableMongoDbMessageStore.java
@@ -15,49 +15,33 @@
  */
 package org.springframework.integration.mongodb.store;
 
+import static org.springframework.integration.mongodb.store.MessageDocumentFields.*;
+
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.Iterator;
+import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.UUID;
 
-import org.springframework.beans.BeansException;
-import org.springframework.beans.DirectFieldAccessor;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.context.ApplicationContext;
-import org.springframework.context.ApplicationContextAware;
-import org.springframework.core.convert.TypeDescriptor;
-import org.springframework.core.convert.converter.Converter;
-import org.springframework.core.convert.converter.GenericConverter;
-import org.springframework.core.serializer.support.DeserializingConverter;
-import org.springframework.core.serializer.support.SerializingConverter;
-import org.springframework.data.annotation.Id;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+
 import org.springframework.data.domain.Sort;
 import org.springframework.data.mongodb.MongoDbFactory;
-import org.springframework.data.mongodb.core.FindAndModifyOptions;
-import org.springframework.data.mongodb.core.IndexOperations;
 import org.springframework.data.mongodb.core.MongoTemplate;
-import org.springframework.data.mongodb.core.convert.CustomConversions;
 import org.springframework.data.mongodb.core.convert.MappingMongoConverter;
-import org.springframework.data.mongodb.core.index.Index;
-import org.springframework.data.mongodb.core.mapping.MongoMappingContext;
 import org.springframework.data.mongodb.core.query.Criteria;
-import org.springframework.data.mongodb.core.query.Order;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.data.mongodb.core.query.Update;
-import org.springframework.integration.IntegrationMessageHeaderAccessor;
-import org.springframework.integration.store.AbstractMessageGroupStore;
-import org.springframework.integration.store.ChannelMessageStore;
 import org.springframework.integration.store.MessageGroup;
 import org.springframework.integration.store.MessageGroupStore;
 import org.springframework.integration.store.MessageStore;
-import org.springframework.integration.store.PriorityCapableChannelMessageStore;
 import org.springframework.integration.store.SimpleMessageGroup;
+import org.springframework.jmx.export.annotation.ManagedAttribute;
 import org.springframework.messaging.Message;
-import org.springframework.messaging.MessageHeaders;
 import org.springframework.util.Assert;
 
 /**
@@ -71,49 +55,16 @@ import org.springframework.util.Assert;
  * @author Artem Bilan
  * @since 3.0
  */
-public class ConfigurableMongoDbMessageStore extends AbstractMessageGroupStore
-		implements MessageStore, ChannelMessageStore, PriorityCapableChannelMessageStore, InitializingBean, ApplicationContextAware {
+public class ConfigurableMongoDbMessageStore extends AbstractConfigurableMongoDbMessageStore
+		implements MessageStore, MessageGroupStore, Iterable<MessageGroup> {
 
 	public final static String DEFAULT_COLLECTION_NAME = "configurableStoreMessages";
 
-	public final static String SEQUENCE_NAME = "messagesSequence";
+	private static final Log logger = LogFactory.getLog(ConfigurableMongoDbMessageStore.class);
 
-	/**
-	 * The name of the message header that stores a flag to indicate that the message has been saved. This is an
-	 * optimization for the put method.
-	 */
-	public static final String SAVED_KEY = ConfigurableMongoDbMessageStore.class.getSimpleName() + ".SAVED";
+	private final Collection<MessageGroupCallback> expiryCallbacks = new LinkedHashSet<MessageGroupCallback>();
 
-	/**
-	 * The name of the message header that stores a timestamp for the time the message was inserted.
-	 */
-	public static final String CREATED_DATE_KEY = ConfigurableMongoDbMessageStore.class.getSimpleName() + ".CREATED_DATE";
-
-	private static final String MESSAGE_ID = "messageId";
-
-	private static final String PRIORITY = "priority";
-
-	private static final String GROUP_ID = "groupId";
-
-	private static final String LAST_MODIFIED_TIME = "lastModifiedTime";
-
-	private static final String SEQUENCE = "sequence";
-
-	private static final String LAST_RELEASED_SEQUENCE = "lastReleasedSequence";
-
-	private static final String COMPLETE = "complete";
-
-	private final String collectionName;
-
-	private final MongoDbFactory mongoDbFactory;
-
-	private volatile MongoTemplate mongoTemplate;
-
-	private volatile MappingMongoConverter mappingMongoConverter;
-
-	private boolean priorityEnabled;
-
-	private ApplicationContext applicationContext;
+	private volatile boolean timeoutOnIdle;
 
 
 	public ConfigurableMongoDbMessageStore(MongoTemplate mongoTemplate) {
@@ -121,11 +72,7 @@ public class ConfigurableMongoDbMessageStore extends AbstractMessageGroupStore
 	}
 
 	public ConfigurableMongoDbMessageStore(MongoTemplate mongoTemplate, String collectionName) {
-		Assert.notNull("'mongoTemplate' must not be null");
-		Assert.hasText("'collectionName' must not be empty");
-		this.collectionName = collectionName;
-		this.mongoTemplate = mongoTemplate;
-		this.mongoDbFactory = null;
+		super(mongoTemplate, collectionName);
 	}
 
 	public ConfigurableMongoDbMessageStore(MongoDbFactory mongoDbFactory) {
@@ -141,64 +88,35 @@ public class ConfigurableMongoDbMessageStore extends AbstractMessageGroupStore
 	}
 
 	public ConfigurableMongoDbMessageStore(MongoDbFactory mongoDbFactory, MappingMongoConverter mappingMongoConverter, String collectionName) {
-		Assert.notNull("'mongoDbFactory' must not be null");
-		Assert.hasText("'collectionName' must not be empty");
-		this.collectionName = collectionName;
-		this.mongoDbFactory = mongoDbFactory;
-		this.mappingMongoConverter = mappingMongoConverter;
+		super(mongoDbFactory, mappingMongoConverter, collectionName);
 	}
 
-	@Override
-	public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
-		this.applicationContext = applicationContext;
-	}
-
-	public void setPriorityEnabled(boolean priorityEnabled) {
-		this.priorityEnabled = priorityEnabled;
-	}
-
-	@Override
-	public boolean isPriorityEnabled() {
-		return this.priorityEnabled;
-	}
-
-	@Override
-	public void afterPropertiesSet() throws Exception {
-		if (this.mongoTemplate == null) {
-			if (this.mappingMongoConverter == null) {
-				this.mappingMongoConverter = new MappingMongoConverter(this.mongoDbFactory, new MongoMappingContext());
-				this.mappingMongoConverter.setApplicationContext(this.applicationContext);
-				List<Object> customConverters = new ArrayList<Object>();
-				customConverters.add(new MongoDbMessageBytesConverter());
-				this.mappingMongoConverter.setCustomConversions(new CustomConversions(customConverters));
-				this.mappingMongoConverter.afterPropertiesSet();
-			}
-			this.mongoTemplate = new MongoTemplate(this.mongoDbFactory, this.mappingMongoConverter);
-			if (this.applicationContext != null) {
-				this.mongoTemplate.setApplicationContext(this.applicationContext);
-			}
+	/**
+	 * Convenient injection point for expiry callbacks in the message store. Each of the callbacks provided will simply
+	 * be registered with the store using {@link #registerMessageGroupExpiryCallback(MessageGroupCallback)}.
+	 *
+	 * @param expiryCallbacks the expiry callbacks to add
+	 */
+	public void setExpiryCallbacks(Collection<MessageGroupCallback> expiryCallbacks) {
+		for (MessageGroupCallback callback : expiryCallbacks) {
+			registerMessageGroupExpiryCallback(callback);
 		}
-
-		IndexOperations indexOperations = this.mongoTemplate.indexOps(this.collectionName);
-		indexOperations.ensureIndex(new Index(MESSAGE_ID, Order.ASCENDING));
-
-		indexOperations.ensureIndex(new Index(GROUP_ID, Order.ASCENDING)
-				.on(LAST_MODIFIED_TIME, Order.DESCENDING)
-				.on(SEQUENCE, Order.DESCENDING));
-
-		indexOperations.ensureIndex(new Index(GROUP_ID, Order.ASCENDING)
-				.on(PRIORITY, Order.DESCENDING)
-				.on(LAST_MODIFIED_TIME, Order.ASCENDING)
-				.on(SEQUENCE, Order.ASCENDING));
 	}
 
+	public boolean isTimeoutOnIdle() {
+		return timeoutOnIdle;
+	}
 
-	@Override
-	public Message<?> getMessage(UUID id) {
-		Assert.notNull(id, "'id' must not be null");
-		MessageDocument document = this.mongoTemplate.findOne(Query.query(Criteria.where(MESSAGE_ID).is(id)),
-				MessageDocument.class, this.collectionName);
-		return (document != null) ? document.getMessage() : null;
+	/**
+	 * Allows you to override the rule for the timeout calculation. Typical timeout is based from the time
+	 * the {@link MessageGroup} was created. If you want the timeout to be based on the time
+	 * the {@link MessageGroup} was idling (e.g., inactive from the last update) invoke this method with 'true'.
+	 * Default is 'false'.
+	 *
+	 * @param timeoutOnIdle The boolean.
+	 */
+	public void setTimeoutOnIdle(boolean timeoutOnIdle) {
+		this.timeoutOnIdle = timeoutOnIdle;
 	}
 
 	@Override
@@ -206,32 +124,6 @@ public class ConfigurableMongoDbMessageStore extends AbstractMessageGroupStore
 		Assert.notNull(message, "'message' must not be null");
 		this.addMessageDocument(new MessageDocument(message));
 		return message;
-	}
-
-	@SuppressWarnings({ "rawtypes", "unchecked" })
-	private void addMessageDocument(MessageDocument document) {
-		Message<?> message = document.getMessage();
-		if (message.getHeaders().containsKey(SAVED_KEY)) {
-			Message<?> saved = getMessage(message.getHeaders().getId());
-			if (saved != null) {
-				if (saved.equals(message)) {
-					return;
-				} // We need to save it under its own id
-			}
-		}
-
-		final long createdDate = document.getCreatedTime() == 0 ? System.currentTimeMillis() : document.getCreatedTime();
-
-		Message<?> result = this.getMessageBuilderFactory().fromMessage(message).setHeader(SAVED_KEY, Boolean.TRUE)
-				.setHeader(CREATED_DATE_KEY, createdDate).build();
-
-		Map innerMap = (Map) new DirectFieldAccessor(result.getHeaders()).getPropertyValue("headers");
-		// using reflection to set ID since it is immutable through MessageHeaders
-		innerMap.put(MessageHeaders.ID, message.getHeaders().get(MessageHeaders.ID));
-		innerMap.put(MessageHeaders.TIMESTAMP, message.getHeaders().get(MessageHeaders.TIMESTAMP));
-
-		document.setCreatedTime(createdDate);
-		this.mongoTemplate.insert(document, this.collectionName);
 	}
 
 	@Override
@@ -249,20 +141,10 @@ public class ConfigurableMongoDbMessageStore extends AbstractMessageGroupStore
 
 
 	@Override
-	public int messageGroupSize(Object groupId) {
-		long lCount = this.mongoTemplate.count(groupIdQuery(groupId), this.collectionName);
-		Assert.isTrue(lCount <= Integer.MAX_VALUE, "Message count is out of Integer's range");
-		return (int) lCount;
-	}
-
-	@Override
 	public MessageGroup getMessageGroup(Object groupId) {
 		Assert.notNull(groupId, "'groupId' must not be null");
-		Sort sort = new Sort(Sort.Direction.DESC, LAST_MODIFIED_TIME, SEQUENCE);
-		if (this.priorityEnabled) {
-			sort = new Sort(Sort.Direction.DESC, PRIORITY).and(sort);
-		}
-		Query query = groupIdQuery(groupId).with(sort);
+
+		Query query = groupIdQuery(groupId).with(new Sort(Sort.Direction.DESC, LAST_MODIFIED_TIME, SEQUENCE));
 		List<MessageDocument> messageDocuments = this.mongoTemplate.find(query, MessageDocument.class,
 				this.collectionName);
 
@@ -294,11 +176,8 @@ public class ConfigurableMongoDbMessageStore extends AbstractMessageGroupStore
 	public MessageGroup addMessageToGroup(Object groupId, Message<?> message) {
 		Assert.notNull(groupId, "'groupId' must not be null");
 		Assert.notNull(message, "'message' must not be null");
-		Sort sort = new Sort(Sort.Direction.DESC, LAST_MODIFIED_TIME, SEQUENCE);
-		if (this.priorityEnabled) {
-			sort = new Sort(Sort.Direction.DESC, PRIORITY).and(sort);
-		}
-		Query query = groupIdQuery(groupId).with(sort);
+
+		Query query = groupIdQuery(groupId).with(new Sort(Sort.Direction.DESC, LAST_MODIFIED_TIME, SEQUENCE));
 		MessageDocument messageDocument = this.mongoTemplate.findOne(query, MessageDocument.class, this.collectionName);
 
 		long createdTime = 0;
@@ -335,35 +214,10 @@ public class ConfigurableMongoDbMessageStore extends AbstractMessageGroupStore
 	}
 
 	@Override
-	public void removeMessageGroup(Object groupId) {
-		this.mongoTemplate.remove(groupIdQuery(groupId), this.collectionName);
-	}
-
-	@Override
-	@SuppressWarnings("rawtypes")
-	public Iterator<MessageGroup> iterator() {
-		Map<Object, MessageGroup> messageGroupMap = new HashMap<Object, MessageGroup>();
-		Query query = Query.query(Criteria.where(GROUP_ID).exists(true));
-		query.fields().include(GROUP_ID);
-		List<Map> groupIds = this.mongoTemplate.find(query, Map.class, this.collectionName);
-		for (Map groupId : groupIds) {
-			Object key = groupId.get(GROUP_ID);
-			if (!messageGroupMap.containsKey(key)) {
-				messageGroupMap.put(key, this.getMessageGroup(groupId));
-			}
-		}
-		return messageGroupMap.values().iterator();
-	}
-
-	@Override
 	public Message<?> pollMessageFromGroup(Object groupId) {
 		Assert.notNull(groupId, "'groupId' must not be null");
 
-		Sort sort = new Sort(LAST_MODIFIED_TIME, SEQUENCE);
-		if (this.priorityEnabled) {
-			sort = new Sort(Sort.Direction.DESC, PRIORITY).and(sort);
-		}
-		Query query = groupIdQuery(groupId).with(sort);
+		Query query = groupIdQuery(groupId).with(new Sort(LAST_MODIFIED_TIME, SEQUENCE));
 		MessageDocument document = this.mongoTemplate.findAndRemove(query, MessageDocument.class, this.collectionName);
 		Message<?> message = null;
 		if (document != null) {
@@ -383,152 +237,94 @@ public class ConfigurableMongoDbMessageStore extends AbstractMessageGroupStore
 		this.updateGroup(groupId, lastModifiedUpdate().set(COMPLETE, true));
 	}
 
+	@Override
+	@SuppressWarnings("rawtypes")
+	public Iterator<MessageGroup> iterator() {
+		Map<Object, MessageGroup> messageGroupMap = new HashMap<Object, MessageGroup>();
+		Query query = Query.query(Criteria.where(GROUP_ID).exists(true));
+		query.fields().include(GROUP_ID);
+		List<Map> groupIds = this.mongoTemplate.find(query, Map.class, this.collectionName);
+		for (Map groupId : groupIds) {
+			Object key = groupId.get(GROUP_ID);
+			if (!messageGroupMap.containsKey(key)) {
+				messageGroupMap.put(key, this.getMessageGroup(groupId));
+			}
+		}
+		return messageGroupMap.values().iterator();
+	}
+
+	@Override
+	public void registerMessageGroupExpiryCallback(MessageGroupCallback callback) {
+		expiryCallbacks.add(callback);
+	}
+
+	@Override
+	public int expireMessageGroups(long timeout) {
+		int count = 0;
+		long threshold = System.currentTimeMillis() - timeout;
+		for (MessageGroup group : this) {
+
+			long timestamp = group.getTimestamp();
+			if (this.isTimeoutOnIdle() && group.getLastModified() > 0) {
+				timestamp = group.getLastModified();
+			}
+
+			if (timestamp <= threshold) {
+				count++;
+				expire(group);
+			}
+		}
+		return count;
+	}
+
+	@Override
+	@ManagedAttribute
+	public int getMessageCountForAllMessageGroups() {
+		int count = 0;
+		for (MessageGroup group : this) {
+			count += group.size();
+		}
+		return count;
+	}
+
+	@Override
+	@ManagedAttribute
+	public int getMessageGroupCount() {
+		int count = 0;
+		for (@SuppressWarnings("unused") MessageGroup group : this) {
+			count ++;
+		}
+		return count;
+	}
+
+	private void expire(MessageGroup group) {
+
+		RuntimeException exception = null;
+
+		for (MessageGroupCallback callback : expiryCallbacks) {
+			try {
+				callback.execute(this, group);
+			} catch (RuntimeException e) {
+				if (exception == null) {
+					exception = e;
+				}
+				logger.error("Exception in expiry callback", e);
+			}
+		}
+
+		if (exception != null) {
+			throw exception;
+		}
+	}
+
 
 	private void updateGroup(Object groupId, Update update) {
-		Sort sort = new Sort(Sort.Direction.DESC, LAST_MODIFIED_TIME, SEQUENCE);
-		if (this.priorityEnabled) {
-			sort = new Sort(Sort.Direction.DESC, PRIORITY).and(sort);
-		}
-		Query query = groupIdQuery(groupId).with(sort);
-		this.mongoTemplate.updateFirst(query, update, this.collectionName);
+		this.mongoTemplate.updateFirst(groupIdQuery(groupId).with(new Sort(LAST_MODIFIED_TIME, SEQUENCE)), update,
+				this.collectionName);
 	}
 
 	private static Update lastModifiedUpdate() {
 		return Update.update(LAST_MODIFIED_TIME, System.currentTimeMillis());
-	}
-
-
-	private static Query groupIdQuery(Object groupId) {
-		return Query.query(Criteria.where(GROUP_ID).is(groupId));
-	}
-
-	private int getNextId() {
-		Query query = Query.query(Criteria.where("_id").is(SEQUENCE_NAME));
-		query.fields().include(SEQUENCE);
-		return (Integer) this.mongoTemplate.findAndModify(query,
-				new Update().inc(SEQUENCE, 1),
-				FindAndModifyOptions.options().returnNew(true).upsert(true),
-				Map.class,
-				this.collectionName).get(SEQUENCE);
-	}
-
-	/**
-	 * The entity class to wrap {@link Message} to the MongoDB document.
-	 */
-	private static class MessageDocument {
-
-		/*
-		 * Needed as a persistence property to suppress 'Cannot determine IsNewStrategy' MappingException
-		 * when the application context is configured with auditing. The document is not
-		 * currently Auditable.
-		 */
-		@SuppressWarnings("unused")
-		@Id
-		private String _id;
-
-		private final Message<?> message;
-
-		@SuppressWarnings("unused")
-		private final UUID messageId;
-
-		@SuppressWarnings("unused")
-		private final Integer priority;
-
-		private Long createdTime = 0L;
-
-		@SuppressWarnings("unused")
-		private Object groupId;
-
-		private Long lastModifiedTime = 0L;
-
-		private Boolean complete = false;
-
-		private Integer lastReleasedSequence = 0;
-
-		@SuppressWarnings("unused")
-		private int sequence;
-
-		public MessageDocument(Message<?> message) {
-			Assert.notNull(message, "'message' must not be null");
-			this.message = message;
-			this.messageId = message.getHeaders().getId();
-			this.priority = new IntegrationMessageHeaderAccessor(message).getPriority();
-		}
-
-		public Message<?> getMessage() {
-			return message;
-		}
-
-		public void setGroupId(Object groupId) {
-			this.groupId = groupId;
-		}
-
-		public Long getLastModifiedTime() {
-			return lastModifiedTime;
-		}
-
-		public void setLastModifiedTime(long lastModifiedTime) {
-			this.lastModifiedTime = lastModifiedTime;
-		}
-
-		public Long getCreatedTime() {
-			return createdTime;
-		}
-
-		public void setCreatedTime(long createdTime) {
-			this.createdTime = createdTime;
-		}
-
-		public Boolean isComplete() {
-			return complete;
-		}
-
-		public void setComplete(boolean complete) {
-			this.complete = complete;
-		}
-
-		public Integer getLastReleasedSequence() {
-			return lastReleasedSequence;
-		}
-
-		public void setLastReleasedSequence(int lastReleasedSequence) {
-			this.lastReleasedSequence = lastReleasedSequence;
-		}
-
-		public void setSequence(int sequence) {
-			this.sequence = sequence;
-		}
-	}
-
-	/**
-	 * A {@link GenericConverter} implementation to convert {@link Message} to
-	 * serialized {@link byte[]} to store {@link Message} to the MongoDB.
-	 * And vice versa - to convert {@link byte[]} from the MongoDB to the {@link Message}.
-	 */
-	private static class MongoDbMessageBytesConverter implements GenericConverter {
-
-		private final Converter<Object, byte[]> serializingConverter = new SerializingConverter();
-
-		private final Converter<byte[], Object> deserializingConverter = new DeserializingConverter();
-
-		@Override
-		public Set<ConvertiblePair> getConvertibleTypes() {
-			Set<ConvertiblePair> convertiblePairs = new HashSet<ConvertiblePair>();
-			convertiblePairs.add(new ConvertiblePair(Message.class, byte[].class));
-			convertiblePairs.add(new ConvertiblePair(byte[].class, Message.class));
-			return convertiblePairs;
-		}
-
-		@Override
-		public Object convert(Object source, TypeDescriptor sourceType, TypeDescriptor targetType) {
-			if (Message.class.isAssignableFrom(sourceType.getObjectType())) {
-				return serializingConverter.convert(source);
-			}
-			else {
-				return deserializingConverter.convert((byte[]) source);
-			}
-		}
-
 	}
 
 }
