+++ spring-integration-mongodb/src/main/java/org/springframework/integration/mongodb/store/MessageDocumentFields.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2014 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.integration.mongodb.store;
+
+/**
+ * @author Artem Bilan
+ *
+ * @since 4.0
+ */
+public final class MessageDocumentFields {
+
+	public static final String MESSAGE_ID = "messageId";
+
+	public static final String PRIORITY = "priority";
+
+	public static final String GROUP_ID = "groupId";
+
+	public static final String LAST_MODIFIED_TIME = "lastModifiedTime";
+
+	public static final String SEQUENCE = "sequence";
+
+	public static final String LAST_RELEASED_SEQUENCE = "lastReleasedSequence";
+
+	public static final String COMPLETE = "complete";
+
+	private MessageDocumentFields() {
+	}
+
+}
