+++ spring-integration-mongodb/src/test/java/org/springframework/integration/mongodb/store/AbstractMongoDbMessageGroupStoreTests.java
@@ -15,26 +15,19 @@
  */
 package org.springframework.integration.mongodb.store;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 import java.util.Collection;
 import java.util.Iterator;
 import java.util.Properties;
 import java.util.UUID;
 
-import org.hamcrest.Matchers;
+import com.mongodb.Mongo;
 import org.junit.Test;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
 import org.springframework.data.mongodb.core.SimpleMongoDbFactory;
-import org.springframework.integration.IntegrationMessageHeaderAccessor;
 import org.springframework.integration.channel.DirectChannel;
-import org.springframework.integration.channel.PriorityChannel;
 import org.springframework.integration.channel.QueueChannel;
 import org.springframework.integration.history.MessageHistory;
 import org.springframework.integration.mongodb.rules.MongoDbAvailable;
@@ -47,8 +40,6 @@ import org.springframework.integration.support.MessageBuilder;
 import org.springframework.messaging.Message;
 import org.springframework.messaging.MessageChannel;
 import org.springframework.messaging.support.GenericMessage;
-
-import com.mongodb.Mongo;
 
 /**
  * @author Oleg Zhurakousky
@@ -502,64 +493,6 @@ public abstract class AbstractMongoDbMessageGroupStoreTests extends MongoDbAvail
 		assertEquals("fooChannel", fooChannelHistory.get("name"));
 		assertEquals("channel", fooChannelHistory.get("type"));
 	}
-
-	protected void testPriorityChannel(String ctx) throws Exception {
-		this.cleanupCollections(new SimpleMongoDbFactory(new Mongo(), "test"));
-		ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext(ctx, this.getClass());
-		context.refresh();
-
-		Object priorityChannel = context.getBean("priorityChannel");
-		assertThat(priorityChannel, Matchers.not(Matchers.instanceOf(PriorityChannel.class)));
-		assertThat(priorityChannel, Matchers.instanceOf(QueueChannel.class));
-
-		QueueChannel channel = (QueueChannel) priorityChannel;
-
-		Message<String> message = MessageBuilder.withPayload("1").setHeader(IntegrationMessageHeaderAccessor.PRIORITY, 1).build();
-		channel.send(message);
-		message = MessageBuilder.withPayload("-1").setHeader(IntegrationMessageHeaderAccessor.PRIORITY, -1).build();
-		channel.send(message);
-		message = MessageBuilder.withPayload("3").setHeader(IntegrationMessageHeaderAccessor.PRIORITY, 3).build();
-		channel.send(message);
-		message = MessageBuilder.withPayload("0").setHeader(IntegrationMessageHeaderAccessor.PRIORITY, 0).build();
-		channel.send(message);
-		message = MessageBuilder.withPayload("2").setHeader(IntegrationMessageHeaderAccessor.PRIORITY, 2).build();
-		channel.send(message);
-		message = MessageBuilder.withPayload("none").build();
-		channel.send(message);
-		message = MessageBuilder.withPayload("31").setHeader(IntegrationMessageHeaderAccessor.PRIORITY, 3).build();
-		channel.send(message);
-
-		Message<?> receive = channel.receive(1000);
-		assertNotNull(receive);
-		assertEquals("3", receive.getPayload());
-
-		receive = channel.receive(1000);
-		assertNotNull(receive);
-		assertEquals("31", receive.getPayload());
-
-		receive = channel.receive(1000);
-		assertNotNull(receive);
-		assertEquals("2", receive.getPayload());
-
-		receive = channel.receive(1000);
-		assertNotNull(receive);
-		assertEquals("1", receive.getPayload());
-
-		receive = channel.receive(1000);
-		assertNotNull(receive);
-		assertEquals("0", receive.getPayload());
-
-		receive = channel.receive(1000);
-		assertNotNull(receive);
-		assertEquals("-1", receive.getPayload());
-
-		receive = channel.receive(1000);
-		assertNotNull(receive);
-		assertEquals("none", receive.getPayload());
-
-		context.close();
-	}
-
 
 	protected abstract MessageGroupStore getMessageGroupStore() throws Exception;
