+++ src/reference/docbook/whats-new.xml
@@ -110,6 +110,16 @@
 				For more information, see <xref linkend="redis-cms"/>.
 			</para>
 		</section>
+		<section id="4.0-priority-channel-mondodb">
+			<title>MongodDB Channel  Message Store</title>
+			<para>
+				MongoDB support now provides a <classname>MongoDbChannelMessageStore</classname> -
+				a <emphasis>channel</emphasis> specific <interfacename>MessageStore</interfacename> implementation.
+				With <code>priorityEnabled = true</code> option it can be used as a reference for
+				<code>&lt;int:priority-queue&gt;</code> to achieve <emphasis>priority</emphasis> order polling for
+				persisted message. For more information see <xref linkend="mongodb-priority-channel-message-store"/>.
+			</para>
+		</section>
 		<section id="4.0-MBeanExport-annotation">
 			<title>@EnableIntegrationMBeanExport</title>
 			<para>
@@ -269,16 +279,6 @@
 				<interfacename>PriorityCapableChannelMessageStore</interfacename>, allowing it to be used as
 				a <code>message-store</code> reference for <code>priority-queue</code>s.
 				For more information, see <xref linkend="jdbc-message-store-channels"/>.
-			</para>
-		</section>
-	</section>
-		<section id="4.0-priority-channel-mondodb">
-			<title>Priority Channel and MongodDB Message Store</title>
-			<para>
-				MongoDB <interfacename>MessageStore</interfacename>s now implement <interfacename>PriorityCapableChannelMessageStore</interfacename>
-				to provide <code>priorityEnabled</code> option and can be used as a reference for <code>&lt;int:priority-queue&gt;</code>
-				to achieve <emphasis>priority</emphasis> order polling for persisted message.
-				For more information see <xref linkend="mongodb-message-store-priority-channel"/>.
 			</para>
 		</section>
 	</section>
